services/gateway-api/requirements.txt
fastapi~=0.110
uvicorn[standard]~=0.29
httpx~=0.27
pydantic-settings~=2.2
pyjwt~=2.9
redis[hiredis]~=5.0
prometheus-client~=0.20
opentelemetry-instrumentation-fastapi~=0.48
opentelemetry-sdk~=1.25
opentelemetry-exporter-otlp~=1.25
-e ../../libs/python/somagent_secrets
